Everywhere: Rename `arm` to `cpu`
