+++ emu/src/lib.rs
@@ -1,6 +1,6 @@
-mod arm;
 mod bitwise;
 pub mod cartridge_header;
+mod cpu;
 pub mod gba;
 mod memory;
 pub mod render;
